chore: refactor type checks
