+++ package.json
@@ -56,7 +56,7 @@
     "bs58": "^4.0.1",
     "buffer-loader": "0.0.1",
     "cids": "~0.5.2",
-    "class-is": "^0.4.0",
+    "class-is": "^0.5.0",
     "ipfs-block": "~0.6.1",
     "is-ipfs": "~0.3.2",
     "multihashes": "~0.4.12",
