+++ src/dag-link/util.js
@@ -2,10 +2,6 @@
 
 const DAGLink = require('./index')
 
-function isDagLink (link) {
-  return link && link.constructor && DAGLink.isDAGLink(link)
-}
-
 function createDagLinkFromB58EncodedHash (link) {
   return new DAGLink(
     link.name ? link.name : link.Name,
@@ -15,5 +11,4 @@ function createDagLinkFromB58EncodedHash (link) {
 }
 
 exports = module.exports
-exports.isDagLink = isDagLink
 exports.createDagLinkFromB58EncodedHash = createDagLinkFromB58EncodedHash
