+++ src/dag-node/addLink.js
@@ -4,7 +4,7 @@ const dagNodeUtil = require('./util')
 const cloneLinks = dagNodeUtil.cloneLinks
 const cloneData = dagNodeUtil.cloneData
 const toDAGLink = dagNodeUtil.toDAGLink
-const DAGLink = require('./../dag-link')
+const DAGLink = require('../dag-link')
 const DAGNode = require('./index')
 const create = require('./create')
 
@@ -12,10 +12,10 @@ function addLink (node, link, callback) {
   const links = cloneLinks(node)
   const data = cloneData(node)
 
-  if ((link.constructor && DAGLink.isDAGLink(link))) {
+  if (DAGLink.isDAGLink(link)) {
     // It's a DAGLink instance
     // no need to do anything
-  } else if (link.constructor && DAGNode.isDAGNode(link)) {
+  } else if (DAGNode.isDAGNode(link)) {
     // It's a DAGNode instance
     // convert to link
     link = toDAGLink(link)
